Fix URL matching regex for git repository
When handling a git repository URL referenced in a supermarket, the
source URL is matched against a regex to extract the revelant parts.

Unfortunately the '.git' suffix, if present, is not correctly stripped
but is always added later. This leads to a URL that can't be cloned.

This changes the regex to correctly strip the '.git' suffix.