Fix bug when generating changelog on all updated cookbooks in a policy
The method all_possible_dep_names of ChefDK::PolicyfileCompiler is a private
method. Therefore, using the --allow-update-all made the application
crash unexpectedly since it tried to use the private method to compute
the list of all dependencies.

We now compute the list from the dependency graph computed by the
compiler.