policyfile: add minimal support for repository with multiple cookbooks
Cookbooks authors may have git repositories holding multiple cookbooks.
This commit adapt the Policyfile git_changelog method to first find the
directory containing the metadata.rb which matches a specific cookbook
then it uses it to compute the changelog.

As there can't be 2 branches or tags with the same name, the versioning
is also different in such "mono-repository" so the git_changelog method
has also been adapted to search for tags/branches prefixed by the name
of the cookbook.
  i.e. <cookbook_name>-1.0.0 or <cookbook_name>-v1.0.0

Cookbooks with metadata.json or different versioning system are not yet
supported.

Change-Id: I5f4b2842caf816f163941f6d27ea926cb903799d